Move things around to have less complex folder structure
