+++ src/MudBlazor.Docs/Pages/Features/Converters/ConvertersPage.razor
@@ -1,4 +1,4 @@
-﻿    @page "/features/converters"
+﻿@page "/features/converters"
 
 <DocsPage>
     <DocsPageHeader Title="Converters" SubTitle="Converters are essential for binding input elements to your model." />
