+++ src/MudBlazor/Components/DatePicker/DateRange.cs
@@ -1,6 +1,5 @@
 ﻿using MudBlazor.Extensions;
 using MudBlazor.Utilities;
-using MudBlazor.Utilities.Converter.Base;
 
 namespace MudBlazor
 {
