+++ src/MudBlazor/Converters/EmptyConverter.cs
@@ -2,7 +2,7 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace MudBlazor.Utilities.Converter;
+namespace MudBlazor;
 
 /// <summary>
 /// A no-op reversible converter that returns inputs unchanged.
