+++ src/MudBlazor/Converters/RangeConverter.cs
@@ -3,8 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Globalization;
+using MudBlazor.Utilities;
 
-namespace MudBlazor.Utilities.Converter;
+namespace MudBlazor;
 
 #nullable enable
 /// <summary>
