+++ src/MudBlazor/Utilities/Exceptions/ConversionException.cs
@@ -2,7 +2,7 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace MudBlazor.Utilities.Converter.Base;
+namespace MudBlazor.Utilities.Exceptions;
 
 #nullable enable
 /// <summary>
